requirements.txt
asgiref==3.2.10
astroid==2.4.2
dj-database-url==0.5.0
Django==3.0.8
django-crispy-forms==1.9.2
django-phonenumber-field==5.0.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumbers==8.12.8
Pillow==7.2.0
psycopg2-binary==2.8.5
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.0.1
wrapt==1.12.1
